update routing to redirect to OTP login instead of customer login
