+++ app/(otp)/customer/register.tsx
@@ -2,25 +2,21 @@ import { View, Text, ScrollView } from 'react-native';
 import { CheckBox } from '@rneui/themed';
 import CustomButton from '@/components/CustomButton';
 import Header from '@/components/login/Header';
-import { Link } from 'expo-router';
+import { Link, router } from 'expo-router';
 import { useState } from 'react';
 import useRegister from '@/hooks/useRegister';
 import Input from '@/components/login/Input';
-import { Provider } from '@/types/Provider';
-import { useRouter } from 'expo-router';
+import { Customer } from '@/types/Customer';
 
 const Register = () => {
   const [checked, setChecked] = useState(false);
 
-  const [provider, setProvider] = useState({
+  const [customer, setCustomer] = useState({
     email: '',
     password: '',
     firstName: '',
     lastName: '',
     confirmPassword: '',
-    businessName: '',
-    businessAddress: '',
-    taxId: '',
   });
 
   const [errorMessages, setErrorMessages] = useState({
@@ -29,82 +25,63 @@ const Register = () => {
     firstName: '',
     lastName: '',
     confirmPassword: '',
-    businessName: '',
-    businessAddress: '',
-    taxId: '',
   });
 
-  const { error, loading, providerRegister } = useRegister();
-  const router = useRouter();
+  const { error, loading, customerRegister } = useRegister();
 
-  const validateForm = (provider: Provider) => {
+  const validateForm = (customer: Customer) => {
     const errors = {
       email: '',
       password: '',
       firstName: '',
       lastName: '',
       confirmPassword: '',
-      businessName: '',
-      businessAddress: '',
-      taxId: '',
     };
 
-    if (!provider.email) {
+    if (!customer.email) {
       errors.email = 'Email is required';
     }
 
-    if (!provider.password || provider.password.length < 8) {
+    if (!customer.password || customer.password.length < 8) {
       errors.password = 'Password must be at least 8 characters';
     }
 
-    if (!provider.confirmPassword) {
+    if (!customer.confirmPassword) {
       errors.confirmPassword = 'Confirm Password is required';
-    } else if (provider.password !== provider.confirmPassword) {
+    } else if (customer.password !== customer.confirmPassword) {
       errors.confirmPassword = 'Passwords do not match';
     }
 
-    if (!provider.firstName) {
+    if (!customer.firstName) {
       errors.firstName = 'First Name is required';
     }
 
-    if (!provider.lastName) {
+    if (!customer.lastName) {
       errors.lastName = 'Last Name is required';
-    }
-
-    if (!provider.businessName) {
-      errors.businessName = 'Business Name is required';
-    }
-
-    if (!provider.businessAddress) {
-      errors.businessAddress = 'Business Address is required';
-    }
-
-    if (!provider.taxId) {
-      errors.taxId = 'Tax ID is required';
     }
 
     return errors;
   };
 
   const handleRegister = async () => {
-    const errors = validateForm(provider);
+    const errors = validateForm(customer);
     setErrorMessages(errors);
 
     const hasErrors = Object.values(errors).some((error) => error.length > 0);
     if (hasErrors) {
       return;
     }
+
     try {
-      const success = await providerRegister(
-        provider.email,
-        provider.password,
-        provider.firstName,
-        provider.lastName,
-        provider.confirmPassword,
-        provider.businessName,
-        provider.businessAddress,
-        provider.taxId,
+      const success = await customerRegister(
+        customer.email,
+        customer.password,
+        customer.firstName,
+        customer.lastName,
+        customer.confirmPassword,
       );
+
+      console.log('success', success);
 
       if (success) {
         router.push('/(otp)/Verification');
@@ -124,9 +101,9 @@ const Register = () => {
             label="First Name"
             placeholder="Enter your first name"
             isPassword={false}
-            value={provider.firstName}
+            value={customer.firstName}
             onChangeText={(text) =>
-              setProvider({ ...provider, firstName: text })
+              setCustomer({ ...customer, firstName: text })
             }
             error={errorMessages.firstName ? true : false}
             errorMessage={
@@ -137,9 +114,9 @@ const Register = () => {
             label="Last Name"
             placeholder="Enter your last name"
             isPassword={false}
-            value={provider.lastName}
+            value={customer.lastName}
             onChangeText={(text) =>
-              setProvider({ ...provider, lastName: text })
+              setCustomer({ ...customer, lastName: text })
             }
             error={errorMessages.lastName ? true : false}
             errorMessage={errorMessages.lastName ? errorMessages.lastName : ''}
@@ -148,8 +125,8 @@ const Register = () => {
             label="Email"
             placeholder="Enter your email"
             isPassword={false}
-            value={provider.email}
-            onChangeText={(text) => setProvider({ ...provider, email: text })}
+            value={customer.email}
+            onChangeText={(text) => setCustomer({ ...customer, email: text })}
             error={errorMessages.email ? true : false}
             errorMessage={errorMessages.email ? errorMessages.email : ''}
           />
@@ -157,9 +134,9 @@ const Register = () => {
             label="Password"
             placeholder="Enter your password"
             isPassword={true}
-            value={provider.password}
+            value={customer.password}
             onChangeText={(text) =>
-              setProvider({ ...provider, password: text })
+              setCustomer({ ...customer, password: text })
             }
             error={errorMessages.password ? true : false}
             errorMessage={errorMessages.password ? errorMessages.password : ''}
@@ -168,49 +145,14 @@ const Register = () => {
             label="Confirm Password"
             placeholder="Re-enter your password"
             isPassword={true}
-            value={provider.confirmPassword}
+            value={customer.confirmPassword}
             onChangeText={(text) =>
-              setProvider({ ...provider, confirmPassword: text })
+              setCustomer({ ...customer, confirmPassword: text })
             }
             error={errorMessages.confirmPassword ? true : false}
             errorMessage={
               errorMessages.confirmPassword ? errorMessages.confirmPassword : ''
             }
-          />
-          <Input
-            label="Business Name"
-            placeholder="Enter your business name"
-            isPassword={false}
-            value={provider.businessName}
-            onChangeText={(text) =>
-              setProvider({ ...provider, businessName: text })
-            }
-            error={errorMessages.businessName ? true : false}
-            errorMessage={
-              errorMessages.businessName ? errorMessages.businessName : ''
-            }
-          />
-          <Input
-            label="Business Address"
-            placeholder="Enter your business address"
-            isPassword={false}
-            value={provider.businessAddress}
-            onChangeText={(text) =>
-              setProvider({ ...provider, businessAddress: text })
-            }
-            error={errorMessages.businessAddress ? true : false}
-            errorMessage={
-              errorMessages.businessAddress ? errorMessages.businessAddress : ''
-            }
-          />
-          <Input
-            label="Tax ID"
-            placeholder="Enter your tax ID"
-            isPassword={false}
-            value={provider.taxId}
-            onChangeText={(text) => setProvider({ ...provider, taxId: text })}
-            error={errorMessages.taxId ? true : false}
-            errorMessage={errorMessages.taxId ? errorMessages.taxId : ''}
           />
         </View>
         <View className="flex-row items-center">
@@ -226,11 +168,13 @@ const Register = () => {
             </Text>
           </Text>
         </View>
+
         {error && (
           <Text className="text-[#FF5757] text-center font-Roboto-Medium text-base mt-3 -mb-3">
             {error}
           </Text>
         )}
+
         <CustomButton
           title="Continue"
           containerStyles="mt-5 bg-[#FDBD10] p-4 rounded-lg w-full shadow-md"
@@ -238,11 +182,12 @@ const Register = () => {
           onPress={handleRegister}
           disabled={loading || !checked}
         />
+
         <View className="flex-row items-center justify-center my-5">
           <Text className="font-Roboto-Light text-black/70 text-base">
             Already have an account?{' '}
             <Link
-              href="/(tabs)/provider/login"
+              href="/(otp)/customer/login"
               className="text-[#147E93] underline font-Roboto-Medium"
             >
               Login
