+++ app/(tabs)/profile/[id]/index.tsx
@@ -38,7 +38,7 @@ export default function Profile() {
           }
         } else {
           setTimeout(() => {
-            router.push('/customer/login');
+            router.push('/(otp)/customer/login');
           }, 100);
         }
       } catch (error) {
